requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.5.30
cffi==1.14.5
charset-normalizer==2.0.3
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2
django-crum==0.7.5
django-import-export==2.8.0
django-widget-tweaks==1.4.5
et-xmlfile==1.1.0
html5lib==1.1
idna==3.2
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.6
Pillow==8.3.1
pycparser==2.20
Pyphen==0.10.0
pytz==2019.3
PyYAML==6.0
requests==2.26.0
six==1.16.0
sqlparse==0.3.1
tablib==3.2.1
tinycss2==1.1.0
tzdata==2022.2
urllib3==1.26.6
WeasyPrint==52.5
webencodings==0.5.1
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
